refactor: filteredRange zustand store로 변경
