+++ packages/view/src/components/TemporalFilter/TemporalFilter.tsx
@@ -6,7 +6,7 @@ import BounceLoader from "react-spinners/BounceLoader";
 import { Button } from "@mui/material";
 
 import { useGlobalData } from "hooks";
-import { useLoadingStore } from "store";
+import { useLoadingStore, useFilteredRangeStore } from "store";
 
 import { filterDataByDate, getMinMaxDate, lineChartTimeFormatter, sortBasedOnCommitNode } from "./TemporalFilter.util";
 import "./TemporalFilter.scss";
@@ -18,8 +18,9 @@ import { createBrush, drawBrush, resetBrush } from "./LineChartBrush";
 import { BRUSH_MARGIN, TEMPORAL_FILTER_LINE_CHART_STYLES } from "./LineChart.const";
 
 const TemporalFilter = () => {
-  const { data, filteredData, setFilteredData, filteredRange, setFilteredRange, setSelectedData } = useGlobalData();
+  const { data, filteredData, setFilteredData, setSelectedData } = useGlobalData();
   const { loading } = useLoadingStore((state) => state);
+  const { filteredRange, setFilteredRange } = useFilteredRangeStore();
   const brushGroupRef = useRef<SVGGElement | null>(null);
   const brushRef = useRef<d3.BrushBehavior<unknown>>();
