+++ packages/view/src/store/index.ts
@@ -1,1 +1,2 @@
 export * from "./loading";
+export * from "./filteredRange";
